Ansible: provision the VM with everything needed to run Craft
This commit introduces an Ansible playbook that installs everything
needed to run Craft CMS. Specifically, it installs PHP7, Apache,
MySQL, and does a few bits of configuration (setting a security key
for Craft, etc). It also installs aptitude on the VM, mainly just
to prevent Ansible showing a warning about it not being available
(without aptitude it works fine, using `apt-get` instead, but it
displays a warning about this every time it runs).

It might be better to split this up into separate files, but for
the purposes of getting it working, I've just added everything to
the one playbook file for now.