Add DB restore shell script
This script can be run at any time by using the `--provision-with`
flag:

    vagrant provision --provision-with=db_restore

It restores the most recent backup in the `storage/backups` folder.

It currently has the default MySQL password hardcoded, so will need
to be updated if the password changes. Ideally it would be better if
this script also used the credentials in the `.env` file (that would
avoid the need to ever touch this file).