Don't try to use NFS on any platforms
This was originally added because NFS supposedly offers some
performance benefits, but isn't supported on Windows. However,
it turns out that it also seems to cause issues on MacOS.

Craft 3 is much more performant than Craft 2 was anyway, so
rather than investigate this issue it seems simpler to just
scrap it for all platforms.